test(upload): move component tests to integration
